create abstract `ProviderInstanceBase` class and refactor existing providers to extend it
